requirements.txt
flask
flask_restful
aiohttp
aiofiles
bs4
googlesearch-python==1.0.1
gitpython
humanize
httpx
heroku3
requests
motor>=3.3.0
opencv-python-headless
pyrogram==2.0.106
pymongo==4.3.2
python-dotenv
pymediainfo
prettytable
git+https://github.com/Itz-fork/py-trans.git
reportlab
svglib
speedtest-cli==2.1.3
tgcrypto 
telegraph
uwuify
